# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "This is the API documentation for the authentication module.",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/shifts": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Retrieve all shifts",
                "description": "Returns a list of all shifts with their related reservations, services, clients, and users.",
                "operationId": "3d46ea27be8c56655adc7e1fa6a4a202",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shifts": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Shift"
                                            }
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Llista de torns"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving shifts",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Create a new shift",
                "description": "Adds a new shift to the system.",
                "operationId": "59c5d592b8719731e025b74e2cc6c5ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "start_time": {
                                        "description": "Start time of the shift",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T08:00:00Z"
                                    },
                                    "end_time": {
                                        "description": "End time of the shift (nullable)",
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2023-01-01T16:00:00Z"
                                    },
                                    "date": {
                                        "description": "Date of the shift",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2023-01-01"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Shift created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shift": {
                                            "$ref": "#/components/schemas/Shift"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Torn creat correctament"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en la validació de dades"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating shift",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/shifts/toggle": {
            "post": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Toggle the active shift",
                "description": "Closes the active shift if one is open, or creates a new shift if none is active.",
                "operationId": "45ad7bab9025cdffacbc80dfac08c159",
                "responses": {
                    "200": {
                        "description": "Shift toggled successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Torn tancat correctament"
                                        },
                                        "shift": {
                                            "$ref": "#/components/schemas/Shift"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "201": {
                        "description": "New shift created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "shift": {
                                            "$ref": "#/components/schemas/Shift"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Torn creat correctament"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error toggling shift",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en gestionar el torn: Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/shifts/status": {
            "get": {
                "tags": [
                    "Shifts"
                ],
                "summary": "Get the current shift status",
                "description": "Checks if there is an active shift (a shift with no end time).",
                "operationId": "ce4daca7ce4116c59baad97d33253a85",
                "responses": {
                    "200": {
                        "description": "Shift status retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "active": {
                                            "description": "True if there is an active shift, false otherwise",
                                            "type": "boolean",
                                            "example": true
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error retrieving shift status",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error en obtenir l'estat del torn: Detailed error message"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Shift": {
                "title": "Shift",
                "description": "Shift model schema",
                "properties": {
                    "id": {
                        "description": "ID of the shift",
                        "type": "integer",
                        "example": 1
                    },
                    "start_time": {
                        "description": "Start time of the shift",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T08:00:00Z"
                    },
                    "end_time": {
                        "description": "End time of the shift",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T16:00:00Z"
                    },
                    "date": {
                        "description": "Date of the shift",
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "created_at": {
                        "description": "Timestamp when the shift was created",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    },
                    "updated_at": {
                        "description": "Timestamp when the shift was last updated",
                        "type": "string",
                        "format": "date-time",
                        "example": "2023-01-01T00:00:00Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}